pyproject.toml
[project]
name = "llmvad"
version = "0.1.0"
description = "LLM-based Video Anomaly Detection"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.50.0",
    "accelerate>=0.20.0",
    "pillow>=9.0.0",
    "numpy>=1.21.0",
    "matplotlib>=3.5.0",
    "opencv-python>=4.5.0",
    "requests>=2.25.0",
    "python-dotenv>=1.0.0",
    "huggingface-hub>=0.20.0",
    "google-generativeai>=0.3.0",
    "google-auth>=2.23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"] 